+++ src/main/java/ru/ITRev/TestProject/dto/FormatFile.java
@@ -11,14 +11,14 @@ public enum FormatFile {
     pdf("pdf"),
     docx("docx");
 
-    private Long id;
+    private Integer id;
     private String value;
 
     FormatFile(String value) {
         this.value = value;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
@@ -32,7 +32,7 @@ public enum FormatFile {
                 .orElseThrow(() -> new BadRequestException("Неизвестный формат файла"));
     }
 
-    public static FormatFile fromId(Long id) {
+    public static FormatFile fromId(Integer id) {
         return Stream.of(FormatFile.values())
                 .filter(x -> x.getId().equals(id)).findFirst()
                 .orElseThrow(() -> new BadRequestException("Неизвестный формат файла"));
